Fix some tests names and autotools v2 tests (#8299)
* fix autotools v2 test

* rename tests

* simplify test

* minor changes

* move to integration
